[GLEAM] Added IRenderTarget::clear(); [SHIBBOLETH] Added explicit to ProxyAllocator constructor. Some size_t fixings. Removed RenderManager from update list.
Added optimization to resource loading pipeline. Split all resource loading into two tasks.
One task to read all files off disk into memory, and another to process the loaded files.

+++ src/Game/Shibboleth_PackFileSystem.cpp
@@ -38,7 +38,7 @@ PackFile::~PackFile(void)
 	}
 }
 
-unsigned int PackFile::size(void) const
+size_t PackFile::size(void) const
 {
 	return _file_size;
 }
